refactor: Change ReturnButton to ts
